Saves session ids from feedback on the server itself
Allows admin users to examine user sessions in more detail to reproduce bugs and
strange behavior.

It stores the information in a DARedis database, and also adds a separate
admin-only interview to directly jump to the session in question

Can also save information on the server even if github isn't set up (though not
optimized for that use yet).