Reflect the logic of sharing answers in question
If you're on a debug server (test or dev) and the admin has enabled sharing,
always share. Adds a note to let users know this is happenning.

The logic is that answers on a test or dev server shouldn't be PII, and when
filling out a bug, it shouldn't be compromising an anonymous position either
(unlike filling out a negative review might be).