Changed session backend to use SQLAlchemy (#18)
In anticipation of scaling better with thousands of feedback items.
The only significant change is to change the GUID to a normal, incrementing SQL int id. GUID isn't a SQL standard thing (only a postgres thing), and tbh we don't get too much from it, so the flexibility of working with other dbs is probably better than keeping it.

Fixes #14.