Check allowed_repo_owners in python, minimal logs
Don't rely on DA YAML to actually enforce not making an issue if it shouldn't,
needs to be in python function, the only thing that runs normally.

Also adds a need block and moves some things around in the send_to_github code,
so we have minimal logs and don't potentially duplicate a lot of actions.

This removes it as allowed_github_users as a interview variable.