Don't pre-load test file
Without this, the following lines start appearing in the logs:

```
Import of docassemble.GithubFeedbackForm.test_feedback_db failed.  ModuleNotFoundError: No module named 'testcontainers'
```

`testcontainers` isn't a runtime dependency, only a test one (can't distinguish between the two until
we get `pyproject.toml` files in all our projects), so it's not needed to be installed on docassemble. The whole
file should be ignored and not loaded anyway.